[Interpreter] Move out the AOT related proxy code.
